requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
motor>=2.5.1
pymongo>=3.12.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.5
dnspython>=2.1.0
email-validator==1.1.3
pydantic==1.8.2
python-dotenv==0.19.0
pillow==8.3.2
aiofiles==0.7.0
httpx==0.19.0
requests>=2.26.0
pytest>=6.2.5
google-generativeai>=0.3.0
sqlalchemy==1.4.23
alembic==1.7.1
psycopg2-binary==2.9.1
redis==3.5.3
prometheus-client==0.11.0
psutil==5.8.0
python-magic==0.4.24